video: Silence warning from compiler.
`SyncTargetBin` and `i` are both `guint`s, which is an unsigned integer
type (Glib-flavoured).  Not sure how big a `guint` can get, but if
`i` ever becomes bigger than `SyncTargetBin`, you're going to get it
wrapping around that maximum value.

So cast it to a signed type.  `gint` seems the logical choice here.

+++ video.c
@@ -289,7 +289,7 @@ gboolean GetVideo(guchar Mode, double Rate, int Skip, gboolean Redraw) {
           Praw += power(fft.out[i]);
 
         for (i=SyncTargetBin-1; i<=SyncTargetBin+1; i++)
-          Psync += power(fft.out[i]) * (1- .5*abs(SyncTargetBin-i));
+          Psync += power(fft.out[i]) * (1- .5*abs((gint)(SyncTargetBin-i)));
 
         Praw  /= (GetBin(2300+CurrentPic.HedrShift, FFTLen) - GetBin(1500+CurrentPic.HedrShift, FFTLen));
         Psync /= 2.0;
